requirements.txt
Flask==1.0.2
gunicorn==19.9.0
line-bot-sdk==1.8.0
requests==2.20.0